fix error as required

+++ python_introduction/finance_calculator.py
@@ -1,7 +1,7 @@
 monthly_income = input("Enter your monthly income:")
 monthly_expense = input("Enter your total monthly expenses:")
 monthly_savings = int(monthly_income) - int(monthly_expense)
-projected_savings = monthly_savings * 12 + (monthly_savings * 12 * 0.05)
+projected_savings = int(monthly_savings * 12 + (monthly_savings * 12 * 0.05))
 
 print("Your monthly savings are $", monthly_savings)
 print("Projected savings after one year, with interest, is: $",projected_savings)
